address llamacloud integration issue and some ui changes
